# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Equipment Platform — isolated development M5
  version: 0.5.0
  description: |
    Local M5 team slice. Production startup is intentionally rejected.
    Protected resources check active workspace membership, effective capabilities and equipment scope.
    Money is an exact SAR decimal string with two fractional digits (0.01–999999999.99).
    Business dates use YYYY-MM-DD independently of audit instants; UI defaults to Asia/Riyadh.
    Unknown request fields are rejected. IDs, ownership, lifecycle and settled totals are server-owned.
    Idempotency-Key is required for creation. Same authorized normalized payload returns the same
    resource with HTTP 200; a changed payload conflicts (409). Scope includes actor/workspace/action.
    Lists contain at most 30 items and have zero-based pages.
    WEB login sets a session-only HttpOnly SameSite=Strict cookie on /api/v1; authenticated writes
    require X-CSRF-Token returned by /auth/me. No browser credential is stored in localStorage.
    NATIVE login returns an opaque accessToken for OS secure storage and Bearer authentication.
    Sessions expire after 8 hours; no refresh token. Logout revokes immediately; re-login is explicit.
    Dev accepts reserved synthetic +966500000000 through +966500000999 numbers, code 123456.
    OTP expires after 5 minutes, permits 5 attempts, resends at most once per 30 seconds and 10/hour.
...
servers:
  - url: http://127.0.0.1:8080/api/v1
security:
  - bearerSession: []
  - cookieSession: []
paths:
...
  /workspaces/{workspace}/selection:
    parameters: [{$ref: '#/components/parameters/Workspace'}]
    put:
      summary: Save last valid workspace preference
      responses: {'200': {description: Selected workspace ID}}
...
  /account/invitations/{id}/accept:
    parameters: [{$ref: '#/components/parameters/Id'}]
    post:
      summary: Atomically accept invitation and activate membership
      responses: {'200': {description: Accepted invitation}, '409': {description: Closed or expired invitation}}
  /account/invitations/{id}/decline:
    parameters: [{$ref: '#/components/parameters/Id'}]
    post:
      summary: Decline invitation
      responses: {'200': {description: Declined invitation}}
  /workspaces/{workspace}/team/invitations:
    parameters: [{$ref: '#/components/parameters/Workspace'}]
    get:
      summary: Workspace invitations, owner only in M5
      responses: {'200': {description: Bounded invitation list}}
    post:
      summary: Invite synthetic phone without sending SMS
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TeamInvite'}}}}
      responses: {'200': {description: Invitation}, '409': {description: Existing active member or pending invitation}}
  /workspaces/{workspace}/team/invitations/{id}/resend:
    parameters: [{$ref: '#/components/parameters/Workspace'}, {$ref: '#/components/parameters/Id'}]
    post:
      summary: Renew pending invitation for seven days
      responses: {'200': {description: Renewed invitation}}
  /workspaces/{workspace}/team/invitations/{id}/cancel:
    parameters: [{$ref: '#/components/parameters/Workspace'}, {$ref: '#/components/parameters/Id'}]
    post:
      summary: Cancel pending invitation
      responses: {'200': {description: Cancelled invitation}}
  /workspaces/{workspace}/team/members:
    parameters: [{$ref: '#/components/parameters/Workspace'}]
    get:
      summary: Active and revoked workspace members
      responses: {'200': {description: Bounded member list}}
  /workspaces/{workspace}/team/members/{user}:
    parameters: [{$ref: '#/components/parameters/Workspace'}, {in: path, name: user, required: true, schema: {type: string, format: uuid}}]
    get:
      summary: Member detail and selected equipment IDs
      responses: {'200': {description: Member}}
    put:
      summary: Owner updates role, capabilities, scope and financial mode; role change resets defaults
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/MemberChange'}}}}
      responses: {'200': {description: Updated member}}
    delete:
      summary: Owner revokes non-owner membership and ends active driver assignment
      responses: {'200': {description: Revoked}}
  /workspaces/{workspace}/drivers/me/assignment:
    parameters: [{$ref: '#/components/parameters/Workspace'}]
    get:
      summary: Current driver assignment or null
      responses: {'200': {description: Equipment ID or null}}
  /workspaces/{workspace}/drivers/eligible:
    parameters: [{$ref: '#/components/parameters/Workspace'}, {in: query, name: equipmentId, required: true, schema: {type: string, format: uuid}}]
    get:
      summary: Up to 100 active driver members eligible for an authorized equipment
      description: Requires DRIVER_ASSIGNMENT_MANAGE and equipment scope. A driver currently assigned outside the caller's scope is excluded. No other team roles or private member fields are returned.
      responses:
        '200': {description: Minimal eligible driver list, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/EligibleDriver'}}}}}
        '404': {$ref: '#/components/responses/Error'}
  /workspaces/{workspace}/drivers/equipment/{equipment}/assignment:
    parameters: [{$ref: '#/components/parameters/Workspace'}, {in: path, name: equipment, required: true, schema: {type: string, format: uuid}}]
    get:
      summary: Current driver for an authorized equipment, or null assignment fields
      description: Requires DRIVER_ASSIGNMENT_MANAGE and equipment scope. Read immediately before replacing a driver.
      responses:
        '200': {description: Current equipment assignment, content: {application/json: {schema: {$ref: '#/components/schemas/CurrentEquipmentAssignment'}}}}
        '404': {$ref: '#/components/responses/Error'}
    delete:
      summary: End active equipment assignment
      responses: {'200': {description: Unassigned}}
  /workspaces/{workspace}/drivers/{user}/assignments:
    parameters: [{$ref: '#/components/parameters/Workspace'}, {in: path, name: user, required: true, schema: {type: string, format: uuid}}]
    get:
      summary: Bounded assignment history
      responses: {'200': {description: Assignment history}}
  /workspaces/{workspace}/drivers/assignments:
    parameters: [{$ref: '#/components/parameters/Workspace'}]
    post:
      summary: Atomically assign or move driver
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [driverUserId, equipmentId], properties: {driverUserId: {type: string, format: uuid}, equipmentId: {type: string, format: uuid}}}}}}
      responses: {'200': {description: Active assignment}}
  /workspaces/{workspace}/financial-submissions:
    parameters: [{$ref: '#/components/parameters/Workspace'}]
    post:
      summary: Create review request; one of amount, note or receipt is required
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SubmissionCreate'}}}}
      responses: {'200': {description: Pending submission}}
  /workspaces/{workspace}/financial-submissions/mine:
    parameters: [{$ref: '#/components/parameters/Workspace'}, {$ref: '#/components/parameters/Page'}]
    get:
      summary: Own submission history, including former assigned equipment
      responses: {'200': {description: Paginated submissions}}
  /workspaces/{workspace}/financial-submissions/review-queue:
    parameters: [{$ref: '#/components/parameters/Workspace'}, {$ref: '#/components/parameters/Page'}]
    get:
      summary: Pending submissions within reviewer equipment scope
      responses: {'200': {description: Paginated pending submissions}}
  /workspaces/{workspace}/financial-submissions/{id}:
    parameters: [{$ref: '#/components/parameters/Workspace'}, {$ref: '#/components/parameters/Id'}]
    get:
      summary: Own submission or scoped reviewer detail
      responses: {'200': {description: Submission detail}}
  /workspaces/{workspace}/financial-submissions/{id}/approve:
    parameters: [{$ref: '#/components/parameters/Workspace'}, {$ref: '#/components/parameters/Id'}]
    post:
      summary: Lock submission and create exactly one M2 expense for original equipment
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EntryCreate'}}}}
      responses: {'200': {description: Approved submission}, '409': {description: Already reviewed}}
  /workspaces/{workspace}/financial-submissions/{id}/reject:
    parameters: [{$ref: '#/components/parameters/Workspace'}, {$ref: '#/components/parameters/Id'}]
    post:
      summary: Reject with required reason; no financial entry created
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [reason], properties: {reason: {type: string, minLength: 1, maxLength: 500}}}}}}
      responses: {'200': {description: Rejected submission}, '409': {description: Already reviewed}}
components:
  securitySchemes:
    bearerSession: {type: http, scheme: bearer}
    cookieSession: {type: apiKey, in: cookie, name: equipment_session}
  parameters:
    Workspace: {in: path, name: workspace, required: true, schema: {type: string, format: uuid}}
    Id: {in: path, name: id, required: true, schema: {type: string, format: uuid}}
    Page: {in: query, name: page, schema: {type: integer, minimum: 0, maximum: 100000, default: 0}}
...
  responses:
    Error:
      description: Stable code and Arabic message. Cross-workspace and missing records both return NOT_FOUND.
      content:
        application/json:
          schema:
            type: object
            required: [code, message]
            properties:
              code: {type: string}
              message: {type: string}
  schemas:
    EligibleDriver:
      type: object
      required: [userId, displayName, currentEquipmentId, currentEquipmentName, startedAt]
      properties:
        userId: {type: string, format: uuid}
        displayName: {type: string}
        currentEquipmentId: {type: [string, 'null'], format: uuid}
        currentEquipmentName: {type: [string, 'null']}
        startedAt: {type: [string, 'null'], format: date-time}
    CurrentEquipmentAssignment:
      type: object
      required: [equipmentId, assignmentId, driverUserId, driverName, startedAt]
      properties:
        equipmentId: {type: string, format: uuid}
        assignmentId: {type: [string, 'null'], format: uuid}
        driverUserId: {type: [string, 'null'], format: uuid}
        driverName: {type: [string, 'null']}
        startedAt: {type: [string, 'null'], format: date-time}
    TeamInvite:
      type: object
      required: [displayName, phone, role]
      properties:
        displayName: {type: string, minLength: 1, maxLength: 100}
        phone: {type: string, description: Synthetic development number normalized to +9665...}
        role: {type: string, enum: [MANAGER, ACCOUNTANT, DRIVER]}
        scope: {type: string, enum: [ALL_EQUIPMENT, SELECTED_EQUIPMENT], description: Omit for DRIVER}
        equipmentIds: {type: array, items: {type: string, format: uuid}}
        capabilities: {type: array, uniqueItems: true, items: {type: string}}
        financialMode: {type: string, enum: [DIRECT, REVIEW]}
    MemberChange:
      type: object
      description: Omitted fields retain current values except role change resets capability defaults.
      properties:
        role: {type: string, enum: [MANAGER, ACCOUNTANT, DRIVER]}
        scope: {type: string, enum: [ALL_EQUIPMENT, SELECTED_EQUIPMENT]}
        equipmentIds: {type: array, items: {type: string, format: uuid}}
        capabilities: {type: array, uniqueItems: true, items: {type: string}}
        financialMode: {type: string, enum: [DIRECT, REVIEW]}
    SubmissionCreate:
      type: object
      description: Driver equipmentId is derived from current active assignment; if supplied it must match.
      properties:
        equipmentId: {type: string, format: uuid}
        amount: {$ref: '#/components/schemas/Money'}
        transactionDate: {type: string, format: date, description: Defaults to today in Asia/Riyadh}
        note: {type: string, maxLength: 1000}
        receipt:
          type: object
          required: [filename, mediaType, base64]
          properties:
            filename: {type: string, maxLength: 200}
            mediaType: {type: string, enum: [image/png, image/jpeg, application/pdf]}
            base64: {type: string, contentEncoding: base64, description: Maximum decoded size 10 MiB}
    EntryCreate:
      type: object
      description: Existing M2 expense creation shape; approval requires SINGLE expense for the submission equipment.
      required: [equipmentId, amount, category, operationDate]
      properties:
        equipmentId: {type: string, format: uuid}
        amount: {$ref: '#/components/schemas/Money'}
        category: {type: string, enum: [FUEL, MAINTENANCE, OTHER]}
        operationDate: {type: string, format: date}
        paidOn: {type: string, format: date}
        note: {type: string}
        paymentStatus: {type: string, enum: [FULL, PARTIAL, UNPAID]}
        initialPaid: {$ref: '#/components/schemas/Money'}
        partyName: {type: string}
        dueDate: {type: string, format: date}
        entryType: {const: EXPENSE}
        expenseScope: {const: SINGLE}
    Money: {type: string, pattern: '^(0|[1-9][0-9]{0,8})\.[0-9]{2}$', example: '350.00'}
...